+++ src/werkzeug/test.py
@@ -354,9 +354,7 @@ class EnvironBuilder:
         self.path = iri_to_uri(request_uri.path)
         self.request_uri = path
         if base_url is not None:
-            base_url = iri_to_uri(
-                base_url, charset=charset if charset != "utf-8" else None
-            )
+            base_url = iri_to_uri(base_url)
         self.base_url = base_url  # type: ignore
         if isinstance(query_string, str):
             self.query_string = query_string